refactor: 拆分应用运行时与前端模块
