+++ templates/index.html
@@ -698,7 +698,11 @@
 <script src="/static/operations-update.js?v={{ app_version }}"></script>
 <script src="/static/operations-center.js?v={{ app_version }}"></script>
 <script src="/static/history-review-center.js?v={{ app_version }}"></script>
+<script src="/static/history-review-notes.js?v={{ app_version }}"></script>
+<script src="/static/history-review-timeline.js?v={{ app_version }}"></script>
 <script src="/static/risk-analysis-center.js?v={{ app_version }}"></script>
+<script src="/static/risk-analysis-render.js?v={{ app_version }}"></script>
+<script src="/static/risk-analysis-comparison.js?v={{ app_version }}"></script>
 <script src="/static/alert-rule-state.js?v={{ app_version }}"></script>
 <script src="/static/alert-rule-socket.js?v={{ app_version }}"></script>
 <script src="/static/alert-rule-list.js?v={{ app_version }}"></script>
@@ -710,6 +714,7 @@
 <script src="/static/alert-configuration-center.js?v={{ app_version }}"></script>
 <script src="/static/portfolio-state.js?v={{ app_version }}"></script>
 <script src="/static/portfolio-render.js?v={{ app_version }}"></script>
+<script src="/static/portfolio-review.js?v={{ app_version }}"></script>
 <script src="/static/portfolio-detail.js?v={{ app_version }}"></script>
 <script src="/static/portfolio-list.js?v={{ app_version }}"></script>
 <script src="/static/portfolio-actions.js?v={{ app_version }}"></script>
