+++ tests/contract_checks.ps1
@@ -312,15 +312,15 @@ Assert-Contains -Path "templates\index.html" -Pattern 'clearRiskHistory' -Messag
 Assert-Contains -Path "templates\index.html" -Pattern 'id="riskCompareButton"' -Message "risk history must expose a comparison action"
 Assert-Contains -Path "templates\index.html" -Pattern 'id="riskComparison"' -Message "risk history must expose a comparison result region"
 Assert-Contains -Path "static\risk-analysis-center.js" -Pattern 'riskComparisonSelection' -Message "risk history comparison must keep transient frontend selection"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern 'function toggleRiskComparisonItem' -Message "risk history must support selecting comparison entries"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern 'function compareSelectedRiskHistory' -Message "risk history must compare selected local entries"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern 'function renderRiskComparison' -Message "risk history must render comparison results"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern '记录一' -Message "risk history comparison must label the first record when timestamps cannot be parsed"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern '记录二' -Message "risk history comparison must label the second record when timestamps cannot be parsed"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern 'data-side-label' -Message "risk history comparison rows must expose their visible side labels"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern 'riskComparisonSelection\.length\s*(>=\s*2|>\s*1)' -Message "risk history comparison must select at most two entries"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern 'function applyRiskHistory[\s\S]*riskComparisonSelection\s*=\s*\[\s*\]' -Message "risk history refresh must reset comparison selection"
-Assert-Contains -Path "static\risk-analysis-center.js" -Pattern '不会调用模型' -Message "risk comparison UI must state that comparison does not call the model"
+Assert-Contains -Path "static\risk-analysis-comparison.js" -Pattern 'function toggleRiskComparisonItem' -Message "risk history must support selecting comparison entries"
+Assert-Contains -Path "static\risk-analysis-comparison.js" -Pattern 'function compareSelectedRiskHistory' -Message "risk history must compare selected local entries"
+Assert-Contains -Path "static\risk-analysis-comparison.js" -Pattern 'function renderRiskComparison' -Message "risk history must render comparison results"
+Assert-Contains -Path "static\risk-analysis-comparison.js" -Pattern '记录一' -Message "risk history comparison must label the first record when timestamps cannot be parsed"
+Assert-Contains -Path "static\risk-analysis-comparison.js" -Pattern '记录二' -Message "risk history comparison must label the second record when timestamps cannot be parsed"
+Assert-Contains -Path "static\risk-analysis-comparison.js" -Pattern 'data-side-label' -Message "risk history comparison rows must expose their visible side labels"
+Assert-Contains -Path "static\risk-analysis-comparison.js" -Pattern 'riskComparisonSelection\.length\s*(>=\s*2|>\s*1)' -Message "risk history comparison must select at most two entries"
+Assert-Contains -Path "static\risk-analysis-render.js" -Pattern 'function applyRiskHistory[\s\S]*riskComparisonSelection\s*=\s*\[\s*\]' -Message "risk history refresh must reset comparison selection"
+Assert-Contains -Path "static\risk-analysis-comparison.js" -Pattern '不会调用模型' -Message "risk comparison UI must state that comparison does not call the model"
 Assert-Contains -Path "static\app.css" -Pattern '\.risk-comparison-row' -Message "risk comparison must expose row layout styles"
 Assert-Contains -Path "static\app.css" -Pattern 'content:attr\(data-side-label\)' -Message "mobile risk comparison rows must render side labels from data attributes"
 Assert-Contains -Path "static\app.css" -Pattern '@media[\s\S]*\.risk-comparison-row' -Message "risk comparison rows must include a mobile layout rule"
@@ -462,14 +462,14 @@ Assert-Contains -Path "goldmonitor\socket_history_review.py" -Pattern '@socketio
 Assert-Contains -Path "goldmonitor\socket_history_review.py" -Pattern '@socketio\.on\("export_review_report"\)' -Message "backend must expose review report export"
 Assert-Contains -Path "static\history-review-center.js" -Pattern 'event_timeline_updated' -Message "frontend must handle event timeline updates"
 Assert-Contains -Path "static\history-review-center.js" -Pattern 'review_report_exported' -Message "frontend must handle review report export results"
-Assert-Contains -Path "static\history-review-center.js" -Pattern 'renderEventTimeline' -Message "frontend must render the event timeline"
+Assert-Contains -Path "static\history-review-timeline.js" -Pattern 'renderEventTimeline' -Message "frontend must render the event timeline"
 Assert-Contains -Path "templates\index.html" -Pattern 'exportReviewReport' -Message "frontend must trigger review report export"
 Assert-Contains -Path "templates\index.html" -Pattern 'createReviewNoteButton' -Message "event timeline must expose review note creation"
 Assert-Contains -Path "templates\index.html" -Pattern 'reviewNoteEditor' -Message "event timeline must expose the review note editor"
 Assert-Contains -Path "templates\index.html" -Pattern 'reviewNoteTitle.*maxlength="80"' -Message "review note title input must enforce the backend limit"
 Assert-Contains -Path "templates\index.html" -Pattern 'reviewNoteContent.*maxlength="2000"' -Message "review note content input must enforce the backend limit"
-Assert-Contains -Path "static\history-review-center.js" -Pattern 'save_review_note' -Message "frontend must submit review note changes"
-Assert-Contains -Path "static\history-review-center.js" -Pattern 'delete_review_note' -Message "frontend must request review note deletion"
+Assert-Contains -Path "static\history-review-notes.js" -Pattern 'save_review_note' -Message "frontend must submit review note changes"
+Assert-Contains -Path "static\history-review-notes.js" -Pattern 'delete_review_note' -Message "frontend must request review note deletion"
 Assert-Contains -Path "static\history-review-center.js" -Pattern 'review_note_saved' -Message "frontend must handle saved review notes"
 Assert-Contains -Path "static\history-review-center.js" -Pattern 'review_note_deleted' -Message "frontend must handle deleted review notes"
 Assert-Contains -Path "static\history-review-center.js" -Pattern 'review_notes_updated' -Message "frontend must refresh the timeline after review note changes"
