+++ tests/test_run_checks_script.py
@@ -13,6 +13,8 @@ EXPECTED_COMPILE_TARGETS = (
     "goldmonitor/alert_rules.py",
     "goldmonitor/daily_digest.py",
     "goldmonitor/config_runtime.py",
+    "goldmonitor/settings_runtime.py",
+    "goldmonitor/history_runtime.py",
     "goldmonitor/desktop_runtime.py",
     "goldmonitor/desktop_status.py",
     "goldmonitor/diagnostics_runtime.py",
@@ -83,6 +85,8 @@ EXPECTED_COMPILE_TARGETS = (
     "tests/test_portfolio_analytics_module.py",
     "tests/test_portfolio_analytics_app.py",
     "tests/test_settings_store_module.py",
+    "tests/test_settings_runtime_module.py",
+    "tests/test_history_runtime_module.py",
     "tests/test_notifications_module.py",
     "tests/test_scheduler_module.py",
     "tests/test_daily_digest_module.py",
@@ -136,6 +140,8 @@ EXPECTED_PYTHON_CHECKS = (
     "tests/test_market_adapters_module.py",
     "tests/test_market_runtime_module.py",
     "tests/test_settings_store_module.py",
+    "tests/test_settings_runtime_module.py",
+    "tests/test_history_runtime_module.py",
     "tests/test_notifications_module.py",
     "tests/test_scheduler_module.py",
     "tests/test_daily_digest_module.py",
